requirements.txt
Flask
python-telegram-bot==20.8
requests
python-binance